+++ jackson/src/main/java/com/twitter/finatra/request/QueryParam.java
@@ -8,4 +8,6 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 @Target({ PARAMETER })
 @Retention(RUNTIME)
-public @interface QueryParam {}
+public @interface QueryParam {
+    String value() default "";
+}