+++ jackson/src/main/java/com/twitter/finatra/request/RouteParam.java
@@ -8,4 +8,6 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 @Target({ PARAMETER })
 @Retention(RUNTIME)
-public @interface RouteParam {}
+public @interface RouteParam {
+    String value() default "";
+}